+++ test/kill_test.rb
@@ -8,6 +8,11 @@ class KillTest < Minitest::Test
 
   def test_kill_long_running_queries
     assert database.kill_long_running_queries
+
+    # not affected by disable_kill option
+    with_disable_kill do
+      assert database.kill_long_running_queries
+    end
   end
 
   def test_kill_all